Safari-611.1.10/LayoutTests/storage/domstorage/localstorage/file-can-access-expected.txt

CONSOLE MESSAGE: PASS: window.localStorage WAS accessible

Test that we can access localStorage from a file URL even if universal access is turned off.
